Updating shell command to build docker images. (#134)
Previous implementation was based on a sequence of two commands - change directory (`cd`) to the build context and then docker build with context directory being equal to the current one. These two commands were separated by a semi-colon `;` and this did not work on windows machines. Moreover, `cd` was redundant anyway since the context directory can be provided to the docker build anyway.