MLBox project for MLBox developers tutorials.
This is an attempt to provide a better tutorial on how to get started with creating MLBoxes. Idea is to provide an dummy MLBox project with all required files present. Developers start with `mlbox.yaml` file. That file contains detailed step by step guide on how to update that file and where to go next.