BugFix: existing python environment (non-system) causes non-zero return code. (#145)
This PR solves the following problem. When SSH runner configures remote python environment, when a non-system python interpreter is specified in a platform file, the following command (pseudo code) is used:
```
[ ! -d "PATH_TO_ROOT_PYTHON_ENVIRONMENT ] && { CREATE_PYTHON_ENVIRONMENT }
```
So, if python environment does not exist, create it. The problem here is that if this environment exists, the condition fails and returns 1. This is the return code that triggers SSH runner to raise an exception.

The following updates are proposed. Add alternative branch (when condition fails) and use more safer if/then/else construct. With the following configuration:
```yaml
interpreter:
    type: "virtualenv"
    python: "python3"
    requirements: "mlcommons-box-docker==0.2.3"
    location: "${HOME}/mlcommons-box/environments"
    name: "mlcommons-box-docker-0.2.1"
```
The following command is generated by the SSH runner:
```shell
if [ ! -d "${HOME}/mlcommons-box/environments/mlcommons-box-docker-0.2.1" ]; then
    mkdir -p ${HOME}/mlcommons-box/environments && cd ${HOME}/mlcommons-box/environments && virtualenv -p python3 mlcommons-box-docker-0.2.1;
else
    true;
fi
```