Bug fix for singularity unit test related to not being able to get singularity version.
The problem is that when running unit tests for SingularityRun (configure/run) with pytest, the `spython` package cannot determine the singularity version - captured output of `singularity --version` command ran with `subprocess` module is empty. This commit introduces the temporary fix by setting the `SPYTHON_SINGULARITY_VERSION` environment variable before running any unit tests (this variable is used internally in `spython` - if it's not empty, it does not run `singularity --version`).