Support for `~` and environment variables in MLCube parameters.
MLCube contract says relative paths in MLCube configuration files are relative with respect to MLCube
workspace directory. In certain cases it makes sense to use absolute paths too. This maybe the case
when we want to reuse host cache directories that many machine learning frameworks use to cache models
and datasets. We also need to be able to resolve `~` (user home directory), as well as environment
variables (BTW, this is probably needs some discussion at some point in time). This environment
variable could be, for instance, `${HOME}`.

This commit:
- Adds support for `~` and environment variables on MLCube parameters.
- Explicitly deals with absolute and relative paths in the code. Previously, the implementation always
  concatenated workspace directory with parameter value. Now it is only done for relative paths.
- Updates documentation with description of absolute and relative paths.
- Not directly related to this feature, documentation now mentions that users can specify access options
  for parameters (which are file system paths) - `ro` (read only) and `rw` (read and write).