Support for auth tokens. (#336)
This commit adds capability to use `auth` tokens when retrieving image manifests from docker registries (for singularity runner). This is done automatically whenever docker registries reply that authentication token is missing. The singularity runner in this case knows the URL endpoint where it needs to request for a token. Instead of requesting anonymous access, the runner will scan for available tokens in a number of docker/singularity/apptainer configuration files. If auth token is found, it will be used as part of authentication request.