Correct usage of build_args for singularity runner.
Existing implementation in some corner cases overrides singularity runner parameters when it should not be doing so. One corner case is the following:

Using docker configuration to build the singularity image file. In this case, the `build_args` parameter is automatically set to `--fakeroot`
whenever it is detected the singularity runtime supports this. This is differet from case when there is singularity section in MLCube configuration
file in which case the `build_args` is correctly read from system settings file (default location is `${HOME}/mlcube.yaml`) or from MLCubec configuration
file.

This fix changes this behavior. The singularity runner will log the warning message whenever it does not override the build_args as it would do previously.